fix(memory): move worker routes into the plugin to satisfy the import-boundary guards
CI's plugin-import-boundary guards failed after moving the memory worker-control
into the plugin: host `runtime/routes/memory-worker-routes.ts` then imported the
plugin's internals, and knip flagged `worker.ts` as unused because the worker
spawn URL still pointed at the old path.

- Move `memory-worker-routes.ts` (and its test) into
  `plugins/defaults/memory/routes/`, alongside the other `memory-*-routes.ts`.
  `runtime/routes/index.ts` aggregates it as a route contribution — the
  sanctioned way host code consumes plugin routes — so there is no host→plugin
  internal import.
- Fix the worker spawn entry URL in the moved worker-control to `./worker.ts`
  (it now lives beside the worker process entry), which also resolves the knip
  unused-file report.
- Refresh both plugin-import-boundary baselines: drop the stale
  `persistence/worker-control.ts` / `../../../persistence/worker-control.js`
  entries and record the moved files' host imports.
