+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -135,6 +135,8 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../../skills/catalog-cache.js",
     "../../../../telemetry/watchdog-events-store.js",
     "../../../../tools/skills/delete-managed.js",
+    "../../../../util/logger.js",
+    "../../../../util/platform.js",
     "../../../channels/types.js",
     "../../../config/loader.js",
     "../../../config/memory-v3-gate.js",
@@ -180,7 +182,6 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../persistence/jobs-store.js",
     "../../../persistence/raw-query.js",
     "../../../persistence/schema/index.js",
-    "../../../persistence/worker-control.js",
     "../../../prompts/persona-resolver.js",
     "../../../prompts/system-prompt.js",
     "../../../prompts/template-detection.js",
@@ -205,6 +206,7 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../util/strip-comment-lines.js",
     "../../../util/truncate.js",
     "../../../util/worker-memory.js",
+    "../../../util/worker-process.js",
     "../../types.js",
     "../injection-presence.js",
     "../injector-order.js",
